requirements.txt
cairocffi==1.7.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chess==1.11.2
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
fonttools==4.58.4
gizeh==0.1.11
graphviz==0.21
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
moviepy==1.0.3
neat-python==0.92
numpy==2.0.2
packaging==25.0
pillow==11.2.1
proglog==0.1.12
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
tqdm==4.67.1
urllib3==2.4.0
zipp==3.23.0
